+++ src/gpu/cache/kernels/lrn_gpu_b8.optimized.cl
@@ -17,7 +17,7 @@ __attribute__((reqd_work_group_size(8, 1, 1)))
 KERNEL (lrn_gpu_b8)(const __global float* input, __global float* output)
 {
     const uint global_id = get_global_id(0);
-    const uint element_offset = get_global_id(1) * (INPUT_BATCH_NUM/8) * INPUT_FEATURE_NUM;
+    const uint element_offset = (uint)get_global_id(1) * (INPUT_BATCH_NUM/8) * INPUT_FEATURE_NUM;
     
     const uint linear_id = global_id + element_offset;
     float8 acc = 0;
